+++ xkcd/Public/Find-XKCD.ps1
@@ -2,7 +2,8 @@ function Find-XKCD {
     <#
     .SYNOPSIS
         Retrieves the details of comics @ https://xkcd.com/ based on whether a specified search string appears
-        in the title text.
+        in the title text (by default). To search for a specified string in the full text of the comic data,
+        use the -FullSearch switch.
 
     .DESCRIPTION
         The Find-XKCD cmdlet creates a local cache of the XKCD API comic data if one is not found to already